moved some headers to types dir
