+++ aether/adapters/proxy.h
@@ -18,7 +18,7 @@
 #define AETHER_ADAPTERS_PROXY_H_
 
 #include "aether/obj/obj.h"
-#include "aether/address.h"
+#include "aether/types/address.h"
 
 namespace ae {
