+++ aether/ae_actions/ping.h
@@ -30,7 +30,7 @@ DISABLE_WARNING_POP()
 #include "aether/server.h"
 #include "aether/channel.h"
 #include "aether/ptr/ptr.h"
-#include "aether/state_machine.h"
+#include "aether/types/state_machine.h"
 #include "aether/actions/action.h"
 #include "aether/actions/action_context.h"
 #include "aether/events/event_subscription.h"
