+++ aether/ae_actions/registration/registration.h
@@ -28,7 +28,7 @@
 #  include "aether/memory.h"
 #  include "aether/ptr/ptr.h"
 #  include "aether/ptr/ptr_view.h"
-#  include "aether/state_machine.h"
+#  include "aether/types/state_machine.h"
 #  include "aether/actions/action.h"
 #  include "aether/actions/action_view.h"
 #  include "aether/events/multi_subscription.h"
