+++ aether/client_connections/client_cloud_connection.h
@@ -22,7 +22,7 @@
 
 #include "aether/uid.h"
 #include "aether/memory.h"
-#include "aether/async_for_loop.h"
+#include "aether/types/async_for_loop.h"
 #include "aether/actions/timer_action.h"
 #include "aether/actions/notify_action.h"
 #include "aether/events/event_subscription.h"
