+++ aether/crypto/sign.h
@@ -33,7 +33,7 @@
 #    include "third_party/libhydrogen/hydrogen.h"
 #  endif
 
-#  include "aether/variant_type.h"
+#  include "aether/types/variant_type.h"
 #  include "aether/reflect/reflect.h"
 
 namespace ae {
