+++ aether/stream_api/safe_stream/safe_stream_types.h
@@ -20,7 +20,7 @@
 #include <cstdint>
 
 #include "aether/common.h"
-#include "aether/ring_buffer.h"
+#include "aether/types/ring_buffer.h"
 #include "aether/transport/data_buffer.h"
 
 namespace ae {
